rename hooks.ts -> api.ts
